pyproject.toml
[project]
name = "dach-mcp-server"
version = "1.0.0"
description = "MCP Server für Gefährdungsbeurteilungen im Dachdecker- & Gerüstbaugewerbe"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.9.0",
    "Pillow>=11.0.0",
    "reportlab>=4.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
dach-mcp = "src.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
